+++ src/lib/components/module/FunctionCard.tsx
@@ -4,22 +4,22 @@ import { Flex, Text } from "@chakra-ui/react";
 import { DotSeparator } from "../DotSeparator";
 import { CustomIcon } from "../icon";
 import { Tooltip } from "lib/components/Tooltip";
+import type { ExposedFunction, Visibility } from "lib/types";
 
 interface FunctionCardProps {
-  isView?: boolean;
-  disabled?: boolean;
-  visibility?: "public" | "friends" | "script";
+  exposedFn: ExposedFunction;
+  onFunctionSelect: (fn: ExposedFunction) => void;
 }
 
-const getIcon = (visibility: FunctionCardProps["visibility"]) => {
+const getIcon = (visibility: Visibility) => {
   switch (visibility) {
-    case "friends":
-      return <CustomIcon name="friends" color="text.dark" boxSize={3} />;
+    case "friend":
+      return <CustomIcon name="friends" color="gray.600" boxSize={3} />;
     case "script":
-      return <CustomIcon name="code" color="text.dark" boxSize={3} />;
+      return <CustomIcon name="code" color="gray.600" boxSize={3} />;
     case "public":
     default:
-      return <CustomIcon name="website" color="text.main" boxSize={3} />;
+      return <CustomIcon name="website" color="gray.600" boxSize={3} />;
   }
 };
 
@@ -28,6 +28,7 @@ const disabledStyle: { [key in `${boolean}`]: FlexProps } = {
     bgColor: "gray.900",
     _hover: { bg: "gray.900" },
     cursor: "not-allowed",
+    pointerEvents: "none",
     borderColor: "gray.700",
   },
   false: {
@@ -39,10 +40,12 @@ const disabledStyle: { [key in `${boolean}`]: FlexProps } = {
 };
 
 export const FunctionCard = ({
-  isView = true,
-  disabled = false,
-  visibility = "public",
+  exposedFn,
+  onFunctionSelect,
 }: FunctionCardProps) => {
+  const { is_entry: isEntry, is_view: isView, visibility, name } = exposedFn;
+  const disabled = !isEntry || visibility !== "public";
+
   return (
     <Flex
       borderRadius={8}
@@ -52,6 +55,7 @@ export const FunctionCard = ({
       flexDirection="column"
       gap={1}
       border="1px solid"
+      onClick={() => onFunctionSelect(exposedFn)}
       {...disabledStyle[String(disabled) as `${boolean}`]}
     >
       <Flex gap={1} justifyContent="space-between" w="full">
@@ -66,28 +70,30 @@ export const FunctionCard = ({
           </Text>
         </Flex>
         <Flex alignItems="center" gap={2}>
-          <Tooltip
-            bg="primary.dark"
-            label="Only functions with “is_entry: true” and “visibility: public” are able to interacted through Celatone’s module interactions."
-          >
-            <Flex>
-              <CustomIcon name="check" color="success.main" boxSize={3} />
-            </Flex>
-          </Tooltip>
-          <DotSeparator bg="gray.600" />
+          {!disabled && (
+            <>
+              <Tooltip
+                bg="primary.dark"
+                label="Only functions with “is_entry: true” and “visibility: public” are interactable through Celatone’s module interactions."
+              >
+                <Flex>
+                  <CustomIcon name="check" color="success.main" boxSize={3} />
+                </Flex>
+              </Tooltip>
+              <DotSeparator bg="gray.600" />
+            </>
+          )}
           <Flex alignItems="center" gap={1}>
             {getIcon(visibility)}
-            <Text
-              variant="body3"
-              color={visibility === "public" ? "text.main" : "text.dark"}
-              textTransform="capitalize"
-            >
+            <Text variant="body3" color="text.dark" textTransform="capitalize">
               {visibility}
             </Text>
           </Flex>
         </Flex>
       </Flex>
-      <Text variant="body2">Function Name</Text>
+      <Text variant="body2" color={disabled ? "text.disabled" : "text.main"}>
+        {name}
+      </Text>
     </Flex>
   );
 };