+++ src/lib/utils/address.ts
@@ -1,4 +1,7 @@
+import { fromBech32, fromHex, toBech32, toHex } from "@cosmjs/encoding";
+
 import type { AddressReturnType } from "lib/app-provider";
+import type { HexAddr, HumanAddr } from "lib/types";
 
 export const getAddressTypeText = (addressType: AddressReturnType) => {
   switch (addressType) {
@@ -12,4 +15,16 @@ export const getAddressTypeText = (addressType: AddressReturnType) => {
     default:
       return "(Invalid Address)";
   }
+};
+
+export const bech32AddressToHex = (addr: HumanAddr): HexAddr =>
+  "0x".concat(toHex(fromBech32(addr).data)) as HexAddr;
+
+export const hexToBech32Address = (
+  prefix: string,
+  hexAddr: HexAddr
+): HumanAddr => {
+  let strip = hexAddr.replace("0x", "");
+  if (strip.length < 40) strip = strip.padStart(40, "0");
+  return toBech32(prefix, fromHex(strip)) as HumanAddr;
 };